Refactoring of test generator routines: streamline code organization
